+++ app/src/main/java/org/lineageos/twelve/fragments/PlaylistFragment.kt
@@ -249,7 +249,7 @@ class PlaylistFragment : Fragment(R.layout.fragment_playlist) {
                     val totalDurationMs = audios.sumOf { audio ->
                         audio?.durationMs ?: 0
                     }
-                    val totalDurationMinutes = totalDurationMs / 1000 / 60
+                    val totalDurationMinutes = (totalDurationMs / 1000 / 60).toInt()
 
                     val tracksCount = resources.getQuantityString(
                         R.plurals.tracks_count,
